fix(migrations): tier_for_balance accepts bigint (was integer)
Migration 0008 failed to apply against a fresh DB:

  ERROR: function public.tier_for_balance(bigint) does not exist
  (SQLSTATE 42883)

The backfill UPDATE at the end of the file calls
tier_for_balance(coalesce(sum(delta), 0)). Because
reps_transactions.delta is integer, sum(delta) returns bigint
(Postgres widens to avoid overflow on aggregation). The function
was declared with parameter type integer, and Postgres won't
implicitly narrow bigint → integer at a call site — so the
function "doesn't exist" in that signature, even though one with
parameter integer is right there.

Changing the parameter to bigint resolves it both for the
sum(delta) call site (matches directly) and for the plpgsql
trigger functions in 0008 + 0009 that pass integer locals
(integer → bigint widens implicitly, no cast needed).

Each migration runs in its own transaction, so 0008 rolled back
cleanly when this errored — no orphaned function or trigger left
behind. After pulling, re-running `npm run db:push` will pick up
from 0008 onward.

+++ supabase/migrations/0008_tier_promotion.sql
@@ -15,8 +15,15 @@
 
 -- ---------------------------------------------------------------- *
 -- Pure function: balance → tier name
+--
+-- Parameter is bigint, not integer: sum(integer) over a column
+-- returns bigint in Postgres, and Postgres won't implicitly narrow
+-- bigint → integer at the call site (only widen integer → bigint).
+-- Using bigint here means callers can pass either type without
+-- explicit casts, both from SQL queries (sum(delta)) and from
+-- plpgsql trigger functions with integer locals.
 -- ---------------------------------------------------------------- *
-create or replace function public.tier_for_balance(balance integer)
+create or replace function public.tier_for_balance(balance bigint)
 returns text
 language sql
 immutable
@@ -29,7 +36,7 @@ as $$
   end;
 $$;
 
-grant execute on function public.tier_for_balance(integer) to authenticated;
+grant execute on function public.tier_for_balance(bigint) to authenticated;
 
 -- ---------------------------------------------------------------- *
 -- Trigger function — recompute the affected member's tier
